Fix NFC The reference to the NFCHelper was being lost, making the NFC read never resolve. The code has been moved to the ViewController, which has a strong reference. I also improved the Text record type parsing, taking into account the payload locale and encoding flags, accepting only text record that begin with HC1. This allows for tags with multiple records and only returning the one containing a DCC.
